feat: 升级项目依赖与架构优化
- **依赖升级**：
    - 升级 Kotlin 至 2.3.20，KSP 至 2.3.6。
    - 升级 Compose Multiplatform 至 1.10.3。
    - 升级 Ktor (3.4.2)、Koin (4.2.1)、Kermit (2.1.0) 等多项核心库版本。
- **构建配置调整**：
    - 移除 `iosX64` 编译目标。
    - 将 Compose 相关依赖从 `compose` 插件引用改为通过 `libs` 版本目录引用。
    - 显式设置 `ktorfit` 编译器插件版本为 `2.3.3`。
- **代码优化与修复**：
    - 替换 `decodeBase64String` 等旧版 Base64 API 为 Kotlin 标准库 `kotlin.io.encoding.Base64`。
    - 在 `Store` 相关逻辑中增加对 `defaultValue` 和 `value` 的非空断言以获取反射类型。
    - 移除 `MainViewModel` 中 `betweenDays` 结果多余的 `toLong()` 转换。
- **Android 平台优化**：
    - `AndroidManifest.xml` 中移除 `READ_EXTERNAL_STORAGE`、`WRITE_EXTERNAL_STORAGE` 等多个不必要的权限申请。
    - 移除了 `gradle.properties` 中针对 `iosArm64` 的 `static` 缓存配置。

+++ composeApp/src/commonMain/kotlin/vip/mystery0/xhu/timetable/viewmodel/MainViewModel.kt
@@ -511,7 +511,7 @@ class MainViewModel : ComposeViewModel() {
                     append(endDateTime.format(thingDateTimeFormatter))
                 }
             }
-            val remainDays = betweenDays(LocalDate.now(), startDateTime.date).toLong()
+            val remainDays = betweenDays(LocalDate.now(), startDateTime.date)
             TodayThingSheet(
                 it.title,
                 it.location,
